requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Unidecode==1.3.8
Werkzeug==3.1.3
WTForms==3.2.1
email_validator
gunicorn
psycopg2-binary
Flask-Login>=0.6.3,<0.7.0
alembic
PyMuPDF
pytesseract
Pillow
openpyxl
python-magic